+++ shared/components/ResourceGraph/ResourceGraph.scss
@@ -20,6 +20,7 @@
   display: flex;
   justify-content: center;
   background: var(--sapNeutralBackground);
+  height: 500px;
 
   & > :first-child {
     width: 100%;
