+++ shared/components/ResourceGraph/buildGraph/index.js
@@ -1,0 +1,13 @@
+import {
+  buildNetworkGraph,
+  networkFlowResources,
+} from './buildNetworkGraph.js';
+import { buildStructuralGraph } from './buildStructuralGraph';
+
+export function buildGraph(data) {
+  const isStructural = !networkFlowResources.includes(
+    data.initialResource.kind,
+  );
+
+  return isStructural ? buildStructuralGraph(data) : buildNetworkGraph(data);
+}
