+++ shared/utils/helpers.js
@@ -1,6 +1,19 @@
 import { randomNamesGenerator } from './randomNamesGenerator/randomNamesGenerator';
 import pluralize from 'pluralize';
 
+export function getApiPath(resourceType, nodes) {
+  const matchedNode = nodes.find(
+    n =>
+      n.resourceType === resourceType || n.navigationContext === resourceType,
+  );
+  try {
+    const url = new URL(matchedNode?.viewUrl);
+    return url.searchParams.get('resourceApiPath');
+  } catch (e) {
+    return null;
+  }
+}
+
 export function findCommonPrefix(initialPrefix, words) {
   if (!words?.length) {
     return initialPrefix;
